Remove shared helpers from tests
These add indirection without improving legibility too much.
