+++ tests/consent.spec.ts
@@ -1,8 +1,4 @@
 import { test, expect, Page } from "@playwright/test";
-import {
-  answerHealthQuestions,
-  answerWhoIsGivingConsent,
-} from "./shared/helpers";
 
 let p: Page;
 
@@ -95,7 +91,13 @@ async function then_i_see_the_health_questions_page() {
 }
 
 async function when_i_answer_the_health_questions() {
-  await answerHealthQuestions(p);
+  const radio = (n: number) =>
+    `input[name="consent_response[question_${n}][response]"][value="no"]`;
+
+  await p.click(radio(0));
+  await p.click(radio(1));
+  await p.click(radio(2));
+  await p.click(radio(3));
 }
 
 async function and_i_triage_the_patient() {
@@ -138,5 +140,7 @@ async function then_i_see_the_consent_form() {
 }
 
 async function when_i_enter_the_consent_details() {
-  await answerWhoIsGivingConsent(p);
+  await p.fill('[name="consent_response[parent_name]"]', "Jane Doe");
+  await p.fill('[name="consent_response[parent_phone]"]', "07412345678");
+  await p.click("text=Mum");
 }
